Tune compute_accelerations hot loop; make dimensions size_t

+++ gsa.hpp
@@ -50,7 +50,7 @@ class GravitationalSearchAlgorithm {
     };
 
     GsaConfig config{};
-    int dimensions{0};
+    size_t dimensions{0};
     std::vector<double> min_bounds;
     std::vector<double> max_bounds;
     ObjectiveFunction objective_fn;
